requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langgraph>=0.0.26
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
groq>=0.4.0 